ai-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pandas==2.2.2
scikit-learn==1.4.2
python-dotenv==1.0.1
